+++ src/quantum/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 
 1.0.0b25
 ++++++++++++++
-* Renamed ``az quantum workspace user create/delete`` to ``add/remove`` and limited both commands to users identified by principal name or object ID.
+* Renamed ``az quantum workspace user create/delete`` to ``add/remove`` and limited both commands to users identified by email address.
 
 1.0.0b24
 +++++++++++++++